dl/coordinator/tests: move simple_file_committer to state_utils
(cherry picked from commit e7c4dbcfb4973365253a3c5663a8753a4260f87f)
